requirements.txt
Flask>=2.0
python-dotenv>=0.21
openai>=0.27.0
